Fixed TierConfigAutomation.update_parameters
