+++ connect/resource/tier_config_automation.py
@@ -6,10 +6,10 @@ Copyright (c) 2019 Ingram Micro. All Rights Reserved.
 """
 from abc import ABCMeta
 
-from typing import Optional
+from typing import Optional, List
 
-from connect.logger import logger
-from connect.models import ActivationTemplateResponse, ActivationTileResponse
+from connect.logger import logger, function_log
+from connect.models import ActivationTemplateResponse, ActivationTileResponse, Param
 from connect.models.exception import FulfillmentFail, FulfillmentInquire, Skip
 from connect.models.tier_config import TierConfigRequest, TierConfigRequestSchema, TierConfig
 from .automation import AutomationResource
@@ -69,3 +69,15 @@ class TierConfigAutomation(AutomationResource):
             return objects[0].configuration
         else:
             return None
+
+    @function_log
+    def update_parameters(self, pk, params):
+        # type: (str, List[Param]) -> str
+        list_dict = []
+        for _ in params:
+            list_dict.append(_.__dict__ if isinstance(_, Param) else _)
+
+        return self.api.put(
+            path=pk,
+            json={'params': list_dict},
+        )
